+++ pentobi/GameModel.h
@@ -193,8 +193,6 @@ public:
 
     Q_INVOKABLE bool save(const QString& file);
 
-    Q_INVOKABLE bool saveAsciiArt(const QString& file);
-
     Q_INVOKABLE void setMoveAnnotation(int moveNumber,
                                        const QString& annotation);