+++ NEWS.md
@@ -7,6 +7,7 @@ Version 24.0-dev (master)
 * Don't use Qt.labs.platform anymore. This replaces the run-time
   dependency on package qml6-module-qt-labs-platform by
   qml6-module-qtcore (on Debian).
+* Removed menu item Export Ascii Art.
 
 Version 23.1-dev (v23-fixes)
 ----------------------------